[#32] refactor: 불필요한 props type 제거 및 폴더 이동
