+++ src/components/Common/PageTitle.tsx
@@ -1,8 +1,7 @@
-import { PageTitleProps } from '@/types/common/PageTitleProps';
 import { Heading } from '@chakra-ui/react';
 import React from 'react';
 
-const PageTitle: React.FC<PageTitleProps> = ({ pageTitle }) => {
+const PageTitle = (pageTitle: string) => {
     return (
         <Heading fontWeight={600} className="mb-4 text-3xl">
             {pageTitle}
