+++ src/components/Common/StateTab.tsx
@@ -1,8 +1,7 @@
-import { StateTabProps } from '@/types/common/StateTabProps';
 import { Flex, Heading } from '@chakra-ui/react';
 import React from 'react';
 
-const StateTab: React.FC<StateTabProps> = ({ tab1, tab2 }) => {
+const StateTab: React.FC<{ tab1: string; tab2: string }> = ({ tab1, tab2 }) => {
     return (
         <Flex className="my-4">
             <Heading
