+++ src/types/common/LikeDislikeProps.ts
