--- src/types/common/PageTitleProps.ts
@@ -1,3 +1,0 @@
-export interface PageTitleProps {
-    pageTitle: string;
-}
