--- src/types/common/StateTabProps.ts
@@ -1,4 +1,0 @@
-export interface StateTabProps {
-    tab1: string;
-    tab2: string;
-}
